cleanup and operator overloads
Bench: 3226011

+++ src/ucioptions.cpp
@@ -25,7 +25,7 @@ void uciOptions::printOptions()
 void uciOptions::uciHash(int value)
 {
     int sizeMB = std::clamp(value, 2, MAXHASH);
-    U64 elements = (static_cast<unsigned long long>(sizeMB) * 1024 * 1024) / sizeof(TEntry);
+    U64 elements = (static_cast<uint64_t>(sizeMB) * 1024 * 1024) / sizeof(TEntry);
     reallocateTT(elements);
 }
